Start including a compiled version in the repo
This makes it possible for users to upgrade with a simple `git pull`,
rather than having to re-download the latest release from GitHub each
time - after checking that there is one, first!

It also converts it from just a `.bblm` file to an entire `.bbprojectd`
file, allowing me to add things like Clippings and Text Filters in the
future. (Planned!)